requirements.txt
streamlit
streamlit-webrtc
opencv-python-headless
Pillow
groq
deep-translator
python-dotenv
av
numpy